// src/lib/supabase/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      work_shifts: {
        Row: {
          bolt: number
          bonuses: number | null
          bonuses_bolt: number | null
          bonuses_glovo: number | null
          bonuses_other: number
          bonuses_stuart: number
          bonuses_uber: number | null
          bonuses_wolt: number | null
          cash_tips_bolt: number
          cash_tips_glovo: number
          cash_tips_other: number
          cash_tips_stuart: number
          cash_tips_uber: number
          cash_tips_wolt: number
          created_at: string
          date: string
          glovo: number
          hours: number
          id: number
          km: number
          orders_bolt: number | null
          orders_glovo: number | null
          orders_other: number
          orders_stuart: number
          orders_uber: number | null
          orders_wolt: number | null
          other_income: number
          other_platform_name: string | null
          stuart: number
          tips: number | null
          tips_bolt: number | null
          tips_glovo: number | null
          tips_other: number
          tips_stuart: number
          tips_uber: number | null
          tips_wolt: number | null
          uber: number
          user_id: string
          wolt: number
        }
        Insert: {
          bolt?: number
          bonuses?: number | null
          bonuses_bolt?: number | null
          bonuses_glovo?: number | null
          bonuses_other?: number
          bonuses_stuart?: number
          bonuses_uber?: number | null
          bonuses_wolt?: number | null
          cash_tips_bolt?: number
          cash_tips_glovo?: number
          cash_tips_other?: number
          cash_tips_stuart?: number
          cash_tips_uber?: number
          cash_tips_wolt?: number
          created_at?: string
          date: string
          glovo?: number
          hours?: number
          id?: never
          km?: number
          orders_bolt?: number | null
          orders_glovo?: number | null
          orders_other?: number
          orders_stuart?: number
          orders_uber?: number | null
          orders_wolt?: number | null
          other_income?: number
          other_platform_name?: string | null
          stuart?: number
          tips?: number | null
          tips_bolt?: number | null
          tips_glovo?: number | null
          tips_other?: number
          tips_stuart?: number
          tips_uber?: number | null
          tips_wolt?: number | null
          uber?: number
          user_id: string
          wolt?: number
        }
        Update: {
          bolt?: number
          bonuses?: number | null
          bonuses_bolt?: number | null
          bonuses_glovo?: number | null
          bonuses_other?: number
          bonuses_stuart?: number
          bonuses_uber?: number | null
          bonuses_wolt?: number | null
          cash_tips_bolt?: number
          cash_tips_glovo?: number
          cash_tips_other?: number
          cash_tips_stuart?: number
          cash_tips_uber?: number
          cash_tips_wolt?: number
          created_at?: string
          date?: string
          glovo?: number
          hours?: number
          id?: never
          km?: number
          orders_bolt?: number | null
          orders_glovo?: number | null
          orders_other?: number
          orders_stuart?: number
          orders_uber?: number | null
          orders_wolt?: number | null
          other_income?: number
          other_platform_name?: string | null
          stuart?: number
          tips?: number | null
          tips_bolt?: number | null
          tips_glovo?: number | null
          tips_other?: number
          tips_stuart?: number
          tips_uber?: number | null
          tips_wolt?: number | null
          uber?: number
          user_id?: string
          wolt?: number
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}